Добавляет webpack и разделяет проект на модули
