+++ src/components/loadMoreButton.js
@@ -1,0 +1,6 @@
+const createLoadMoreButton = () => {
+  return (`<button class="load-more" type="button">load more</button>`
+  );
+};
+
+export default createLoadMoreButton;
